cut the helper function into two because jetson library do not need to test
