+++ .travis.yml
@@ -1,0 +1,5 @@
+language: python
+python:
+  - "3.6"
+script:
+  - python code/coordinate_planning/testing.py
