+++ code/coordinate_planning/helper_function.py
@@ -1,4 +1,3 @@
-from jetbot import Robot
 import math
 dis_threshold = 0.1
 def check_angle(angle):     #use to convert the angle to be between -180 to 180 degree
@@ -46,19 +45,3 @@ def convert_angle(delta_x,delta_y):
             return math.degrees(math.atan(temp)) + 180 - 360
 
 
-#use for basic control of the robot
-robot = Robot()
-def turn_left(leftspeed = 0, rightspeed = 0.6):    # use for easy control of the car
-    robot.set_motors(leftspeed, rightspeed)
-
-def turn_right(leftspeed = 0.6, rightspeed = 0):    
-    robot.set_motors(leftspeed, rightspeed)
-
-def spin_left(speed = 0.5):
-    robot.left(speed)
-
-def spin_right(speed = 0.5):
-    robot.right(speed)
-
-def straight(speed = 0.3):
-    robot.forward(speed)
